[ALPS03523955] scp: fix 32bit kernel build error
[Detail]
fix 32bit kernel build error and revise 6775 emi include header

MTK-Commit-Id: c32a60fc5c7758591e1c6b0813186154d0782ddb

CR-Id: ALPS03523955
Feature: TinySys

+++ drivers/misc/mediatek/scp/mt6771/scp_awake.c
@@ -245,7 +245,7 @@ int scp_sys_full_reset(void)
 
 	/*copy loader to scp sram*/
 	pr_debug("[SCP]copy to sram\n");
-	memcpy_to_scp(SCP_TCM, (const void *)scp_loader_base_virt, scp_loader_size);
+	memcpy_to_scp(SCP_TCM, (const void *)(size_t)scp_loader_base_virt, scp_loader_size);
 	/*set info to sram*/
 	pr_debug("[SCP]set firmware info to sram\n");
 	writel(scp_fw_base_phys, SCP_TCM + 0x408);
